Uncouple door state information from config.Door type
This commit moves door state information (such as current state, LastCmdTimestamp, etc) back into the door.Door type.

The reason this was done in the first place was to leverage a single type (config.Door) for unmarshaling API responses in the Porter client library (which clearly wasn't a great idea).

To remedy this, a new type, V1DoorState, has been defined in the Porter API client package. This type contains all the configurable properties of a config.Door alongside the state information provided by door.Door in API calls.